Modifica cancel itinerary

+++ main/business_logic/Curator.java
@@ -3,7 +3,6 @@ package main.business_logic;
 import main.DomainModel.*;
 import main.orm.ArtworkDAO;
 import main.orm.ItineraryDAO;
-import main.orm.VisitDAO;
 import main.orm.VisitorDAO;
 
 import javax.mail.*;
@@ -28,7 +27,7 @@ public class Curator {
     public void addArtwork(Itinerary i, Artwork a) throws SQLException, MessagingException {
         //se opera non ancora presente nel db viene memorizzata
         ArtworkDAO adao = new ArtworkDAO();
-        if(adao.get(a.getCode()) == null){
+        if (adao.get(a.getCode()) == null) {
             adao.insert(a);
         }
         ItineraryDAO idao = new ItineraryDAO();
@@ -40,7 +39,7 @@ public class Curator {
         ArrayList<Visitor> nlsubscribers = vdao.getNLSubscribers();
 
         //invio email
-        if (nlsubscribers.size()>0){
+        if (nlsubscribers.size() > 0) {
             Properties properties = new Properties();
             properties.put("mail.smtp.auth", "true");  //autenticazione user
             properties.put("mail.smtp.host", "smtp.gmail.com");  //server smtp gmail
@@ -57,9 +56,9 @@ public class Curator {
             Message message = new MimeMessage(session);
             message.setFrom(new InternetAddress(emailAddress));
             message.setSubject("Nuova opera");
-            message.setText("Nuova opera "+a.getName() +" dell'autore " + a.getAuthor() + " con stato " + a.getStatus() + " presente nell'itinerario " + i.getName());
+            message.setText("Nuova opera " + a.getName() + " dell'autore " + a.getAuthor() + " con stato " + a.getStatus() + " presente nell'itinerario " + i.getName());
 
-            for(Visitor subscriber : nlsubscribers){
+            for (Visitor subscriber : nlsubscribers) {
                 Address addressTo = new InternetAddress(subscriber.getEmailAddress());
                 message.addRecipient(Message.RecipientType.TO, addressTo);
             }
@@ -67,7 +66,7 @@ public class Curator {
         }
     }
 
-    public ArrayList<Artwork> viewArtworks() throws SQLException{
+    public ArrayList<Artwork> viewArtworks() throws SQLException {
         ArtworkDAO adao = new ArtworkDAO();
         return adao.getAll();
     }
@@ -76,14 +75,14 @@ public class Curator {
         //se lo stato è lo stesso, nessun cambiamento
         //se lo stato da visibile passa a non visibile, notificati i visitatori interessati dalla modifica
         //se lo stato da non visibile passa a visibile notificati i newsletter subscriber
-        if(!Objects.equals(as.getStatus(), a.getStatus())){
-            if(a.getArtworkStatusObject().getClass() == OnDisplay.class){
+        if (!Objects.equals(as.getStatus(), a.getStatus())) {
+            if (a.getArtworkStatusObject().getClass() == OnDisplay.class) {
                 //opera passa ad uno stato non visibile
                 VisitorDAO dao = new VisitorDAO();
                 //metodo transitive che fornisce visitatori interessati
                 ArrayList<Visitor> visitors = dao.getToBeNotifiedVisitors(a);
                 //emailing ai visitatori
-                if (visitors.size()>0){
+                if (visitors.size() > 0) {
                     Properties properties = new Properties();
                     properties.put("mail.smtp.auth", "true");  //autenticazione user
                     properties.put("mail.smtp.host", "smtp.gmail.com");  //server smtp gmail
@@ -100,21 +99,20 @@ public class Curator {
                     Message message = new MimeMessage(session);
                     message.setFrom(new InternetAddress(emailAddress));
                     message.setSubject(a.getName() + " non più visibile");
-                    message.setText("L'opera "+ a.getName() +" dell'autore " + a.getAuthor() + " è " + a.getStatus());
+                    message.setText("L'opera " + a.getName() + " dell'autore " + a.getAuthor() + " è " + a.getStatus());
 
-                    for(Visitor visitor : visitors){
+                    for (Visitor visitor : visitors) {
                         Address addressTo = new InternetAddress(visitor.getEmailAddress());
                         message.addRecipient(Message.RecipientType.TO, addressTo);
                     }
                     Transport.send(message);
                 }
-            }
-            else if(as.getClass() == OnDisplay.class){
+            } else if (as.getClass() == OnDisplay.class) {
                 //opera torna allo stato visibile
                 VisitorDAO dao = new VisitorDAO();
                 ArrayList<Visitor> nlsubscribers = dao.getNLSubscribers();
                 //invio email
-                if (nlsubscribers.size()>0) {
+                if (nlsubscribers.size() > 0) {
                     Properties properties = new Properties();
                     properties.put("mail.smtp.auth", "true");  //autenticazione user
                     properties.put("mail.smtp.host", "smtp.gmail.com");  //server smtp gmail
@@ -146,8 +144,11 @@ public class Curator {
         }
     }
 
-    public void cancelItinerary(Itinerary i) throws SQLException{
+    public void cancelItinerary(Itinerary i) throws SQLException {
         ItineraryDAO dao = new ItineraryDAO();
-        dao.delete(i);
+        if (!dao.inUse(i))
+            dao.delete(i);
+        else
+            throw new RuntimeException("Itinerary in use");
     }
 }
